// nnc-web/web/src/interfaces/core-response/IDataset.ts
/* Copyright 2024 Sony Group Corporation. */
/**
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
*/

export interface ICreateDataset {
  upload_url: string;
  encrypted_text: string;
}

interface Metadata {
  offset: number;
  limit: number;
  sort_by: string;
  total: number;
}

interface IFeatures {
  shape?: number[];
  type: string;
  variable: string;
}

export interface IDataset {
  column_num: number;
  csv_header: string;
  data_num: number;
  dataset_id: string;
  dataset_name: string;
  extract_progress: number;
  owner_user_id: string;
  status: string;
  storage_used: number;
  tenant_id: string;
  update_datetime: string;
  labels: string[];
  features: IFeatures[];
}

export interface ISampleDatasetResponse {
  metadata: Metadata;
  datasets: IDataset[];
}

export interface IDatasetResponse {
  metadata: Metadata;
  datasets: IDataset[];
}